Add DeleteAsync method to OffersController (#52)
* Add DeleteAsync method to OffersController

A new DeleteAsync method has been added to the OffersController class in the MeetApp.Backend.Controllers.Api.V1 namespace. This method is an HTTP DELETE endpoint that allows anonymous users to delete an offer by its id.

The method is decorated with several attributes:
- [AllowAnonymous] allows the endpoint to be accessed without authentication.
- [HttpDelete("{id}")] specifies that this is an HTTP DELETE request and the id parameter is taken from the route.
- [Produces(MediaTypeNames.Application.Json)] indicates that the response will be in JSON format.
- [ProducesResponseType(StatusCodes.Status200OK)] indicates that a 200 OK status will be returned if the deletion is successful.
- [ProducesResponseType(StatusCodes.Status404NotFound)] indicates that a 404 Not Found status will be returned if the offer with the specified id does not exist.

The method retrieves the offer from the database using the provided id. If the offer is found, it is removed from the database and a 200 OK response is returned. If the offer is not found, a 404 Not Found response is returned.

* Refactor: Adjust formatting and reorder using directives

Minor formatting change in OffersController.cs by adding an extra closing brace.
Reordered using directives in 20241016172043_Offers_CREATE.cs and Offer.cs for consistency and potential dependency resolution.